requirements.txt
fastapi==0.104.1
pymongo==4.6.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pydantic==2.4.2
uvicorn==0.24.0
python-multipart==0.0.6
bcrypt==4.0.1
motor==3.3.2 